Improve error handling when response is not JSON
The GraphAPI class currently handles HTTP response errors by treating
the response body as JSON and attempting to decode it. This is usually
fine, but there are some cases where Facebook returns responses that do
not contain valid JSON. In these cases a rather uninformative ValueError
is raised indicating that JSON decoding failed. Since we have the HTTP
response at the point where this error is raised, the status code and
body can be extracted to produce a more helpful error message.

This commit makes the following changes:

- Factor out duplicate error handling code into new method
  GraphAPI.create_exception_for_error.

- Add logic to catch JSON decoding error and produce a more informative
  message containing the contents of the HTTP response.

- Add exception class GraphAPIResponseError that is raised when JSON
  decoding fails.

- Add exception base class FacebookError and have both GraphAPIError and
  GraphAPIResponseError inherit from it.