Fix broken request error handling
The code for catching and handling HTTPError was originally written
against urllib2. When the change was made to requests this code was not
properly modified. Unfortunately, this had the effect of completely
suppressing exceptions on error responses rather than handling them.
Fixing this revealed 2 unit-tests that should actually have been failing
the entire time. These tests have now been skipped pending a rewrite.