Current docs don't show how to get an access token
I had to pull this from the unit tests to figure out how to get and use access token cleanly. The current example is not enough, and fumbling through the docs/code first made me think that this was the only way to go about it, which is gross/wrong:

```
    graph = facebook.GraphAPI(
        facebook.GraphAPI().get_app_access_token(
            'your_app_id', 'your_app_secret'
        ),
        version='2.2'
    )
```

Please either merge this or update the docs so that someone can get a working example from just this first example.